chore: configured Intent Extension, fetchTextMessage, in IntentExtensions.kt. Deprecated fun-processIntentAndFetchTextMessages.
