+++ app/src/main/java/kib/project/fast/ui/component/SystemBroadcastReceiver.kt
@@ -9,11 +9,14 @@ import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberUpdatedState
 
+/*
+* Generic Composable to register/unregister broadcast receivers & handle onReceive()
+* */
 @Composable
 fun SystemBroadcastReceiver(
     context: Context,
     systemAction: String, // Sample: Telephony.Sms.Intents.SMS_RECEIVED_ACTION
-    onSystemEvent: (intent: Intent?) -> Unit
+    onSystemEvent: (intent: Intent) -> Unit
 ) {
     // Safely use the latest onSystemEvent lambda passed to the function
     val currentOnSystemEvent by rememberUpdatedState(onSystemEvent)
@@ -23,7 +26,9 @@ fun SystemBroadcastReceiver(
         val intentFilter = IntentFilter(systemAction)
         val broadcast = object : BroadcastReceiver() {
             override fun onReceive(context: Context?, intent: Intent?) {
-                currentOnSystemEvent(intent)
+                intent?.let {
+                    currentOnSystemEvent(it)
+                }
             }
         }
