+++ app/src/main/java/kib/project/fast/ui/component/viewmodels/MpesaTextMessageViewModel.kt
@@ -23,13 +23,4 @@ class MpesaTextMessageViewModel: ViewModel() {
             it.copy(mpesaMessage = message)
         }
     }
-
-    init {
-        //
-    }
-
-    override fun onCleared() {
-        //
-        super.onCleared()
-    }
 }
